refactor: message 디렉터리 안에 sync, async 를 나누어 저장
