+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/message/sync/purchase/controller/PurchaseServiceController.java
@@ -1,6 +1,6 @@
-package com.blockpage.memberservice.adaptor.infrastructure.external.purchase.controller;
+package com.blockpage.memberservice.adaptor.infrastructure.message.sync.purchase.controller;
 
-import com.blockpage.memberservice.adaptor.infrastructure.external.purchase.requestbody.RequestPurchase;
+import com.blockpage.memberservice.adaptor.infrastructure.message.sync.purchase.requestbody.RequestPurchase;
 import com.blockpage.memberservice.application.port.out.port.PurchasePort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
