+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/mysql/persistance/MemberAdaptor.java
@@ -1,10 +1,10 @@
-package com.blockpage.memberservice.adaptor.infrastructure.persistance;
+package com.blockpage.memberservice.adaptor.infrastructure.mysql.persistance;
 
 import static com.blockpage.memberservice.exception.ErrorCode.NICKNAME_ALREADY_EXIST;
 
-import com.blockpage.memberservice.adaptor.infrastructure.entity.MemberEntity;
-import com.blockpage.memberservice.adaptor.infrastructure.repository.MemberRepository;
-import com.blockpage.memberservice.adaptor.infrastructure.value.Role;
+import com.blockpage.memberservice.adaptor.infrastructure.mysql.entity.MemberEntity;
+import com.blockpage.memberservice.adaptor.infrastructure.mysql.value.Role;
+import com.blockpage.memberservice.adaptor.infrastructure.mysql.repository.MemberRepository;
 import com.blockpage.memberservice.application.port.out.port.MemberPort;
 import com.blockpage.memberservice.domain.Member;
 import com.blockpage.memberservice.exception.CustomException;
