+++ src/main/java/com/blockpage/memberservice/adaptor/infrastructure/mysql/repository/RatingRepository.java
@@ -1,6 +1,6 @@
-package com.blockpage.memberservice.adaptor.infrastructure.repository;
+package com.blockpage.memberservice.adaptor.infrastructure.mysql.repository;
 
-import com.blockpage.memberservice.adaptor.infrastructure.entity.RatingEntity;
+import com.blockpage.memberservice.adaptor.infrastructure.mysql.entity.RatingEntity;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
